+++ runner.rb
@@ -1,0 +1,7 @@
+require './app'
+require './text'
+
+class Menu
+  app = App.new
+  app.start
+end
